requirements.txt
django
psycopg2-binary
pillow
ipython
tzdata
pytilsasgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
django-crispy-forms==2.1
executing==2.0.0
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg2==2.9.9
pure-eval==0.2.2
Pygments==2.16.1
pytils==0.4.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.11.2
wcwidth==0.2.8